# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: eSPX Management JSON API
  version: 1.0.0
  description: |
    Machine-readable contract for `/api/v1` JSON routes served by `cmd/management`
    (`internal/management` + `internal/adminapi`). HTMX `/admin/*` HTML routes are excluded.
    HTTP 501 stub routes (deferred dashboards/reports) are excluded — see `internal/adminapi/stub_routes.go` and `internal/openapi/stubs.go`.
servers:
  - url: http://127.0.0.1:8188
    description: Local management (default SERVER_PORT)
components:
  securitySchemes:
    AdminAPIKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Shared admin API key (`ADMIN_API_KEY`).
    SessionCookie:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT session cookie set by `POST /api/v1/auth/login`.
    ConsentHMAC:
      type: apiKey
      in: header
      name: X-Consent-Signature
      description: HMAC-SHA256 of raw request body (`CONSENT_HMAC_SECRET`).
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    GenericObject:
      type: object
      additionalProperties: true
...
paths:
...
  /api/v1/customers/{id}/quota-bump:
    post:
      tags: [Customers]
      summary: POST /api/v1/customers/{id}/quota-bump
      operationId: post_customers_id_quota_bump
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/{id}/quota-status:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/quota-status
      operationId: get_customers_id_quota_status
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/{id}/subscription:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/subscription
      operationId: get_customers_id_subscription
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Customers]
      summary: POST /api/v1/customers/{id}/subscription
      operationId: post_customers_id_subscription
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/customers/{id}/usage:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/usage
      operationId: get_customers_id_usage
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/customers/{id}/usage/daily:
    get:
      tags: [Customers]
      summary: GET /api/v1/customers/{id}/usage/daily
      operationId: get_customers_id_usage_daily
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/license/status:
    get:
      tags: [Customers]
      summary: GET /api/v1/license/status
      operationId: get_license_status
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/selfserve/usage:
    get:
      tags: [Customers]
      summary: GET /api/v1/selfserve/usage
      operationId: get_selfserve_usage
      security:
        - AdminAPIKey: []
        - SessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
...